Make each type's id individual
The id of a type is basically the dataPath of the value it is
representing. But this only ensures uniqueness of the ids if we have one
data object, where each dataPath can only occur once. Different
instances of Fui or different windows can have different data objects
causing naming conflicts. So we prefix each type's id with its
schemaRoot's id. A schemaRoot is either a instance of Fui oder a window
or any other future component that may use a different data object.